Kalman Filters Documentation

Unscented Kalman Filter

UnscentedKalmanFilter.UKFState — Type

UKFState(
    ;dim_x::Int64, dim_z::Int64, σ_parameters::MerweScaled,
    add_x::Function = default_add_x,
    residual_x::Function = default_residual,
    residual_z::Function = default_residual,
    mean_x::Function = default_mean_f,
    mean_z::Function = default_mean_f
)

State of the Unscented Kalman Filter.

Parameters

dim_x::Int64: State dimension

dim_z::Int64: Measurement dimension

x::Array{Float64, 1}: State of the filter

P::Array{Float64, 2}: Covariance matrix of the filter.

K::Array{Float64, 2}: Kalman gain of the filter.

S::Array{Float64, 2}: System uncertainty.

S_inv::Array{Float64, 2}: Inverse of system uncertainty.

σ_parameters::MerweScaled: Parameters for Van der Merwe sigma points.

σ_fx::Array{Float64, 2}: Sigma points after fx.

σ_hx::Array{Float64, 2}: Sigma points after hx.

add_x::Function: Function to calculate addition between two states.

residual_x::Function: Function to calculate residual between two states.

residual_z::Function: Function to calculate residual between state and measurement.

mean_x::Function: Function to calculate mean between Σ-points and weights after passing them through fx.

mean_z::Function: Function to calculate mean between Σ-points and weights after passing them through hx.

UnscentedKalmanFilter.predict! — Function

predict!(
    ;ukf::UKFState, δt::Float64, Q::Union{Nothing, Array{Float64, 2}},
    fx::Function, fx_args...
)

Perform prediction of the Unscented Kalman Filter (UKF). This updates x and P of the ukf to contain predicted state and covariance.

Arguments

ukf::UKFState: State of the UKF on which to perform prediction.

δt::Float64: Time delta.

Q::Union{Nothing, Array{Float64, 2}}: Process noise.

fx::Function(;x::Array{Float64, 1}, δt::Float64, ...): State transition function. Should accept keyword-only arguments, with mandatory x and δt. All other arguments may serve as control input or provide additional info.

fx_args...: Arguments to pass to fx function. Must contain at least x and δt arguments.

UnscentedKalmanFilter.update! — Function

update!(
    ;ukf::UKFState, z::Array{Float64, 1}, R::Union{Nothing, Array{Float64, 2}},
    hx::Function, hx_args...
)

Perform update step of the Unscented Kalman Filter (UKF) given measurements. This updates x and P of the ukf_state to contain updated state and covariance.

Arguments

ukf::UKFState: State of the UKF on which to perform update step.

z::Array{Float64, 1}: Measurements.

R::Union{Nothing, Array{Float64, 2}}: Measurement noise.

hx::Function: Measurement function. Should accept keyword-only arguments with mandatory x argument for state. Given state x of UKF it must transform it to the respective measurement.

hx_args...: Arguments that will be passed to hx function. Should contain at least x argument.

SigmaPoints

SigmaPoints.MerweScaled — Type

MerweScaled(
    ;n::Int64, α::Float64, β::Float64, κ::Float64, residual_x::Function = -
)

Parameters for the Van der Merwe's σ-points

n::Int64: Dimensionality of the state. 2n+1 will be generated

α::Float64: Spread of the σ-points around the mean.

β::Float64: Prior knowledge of the distribution of the mean.

κ::Float64: Second scaling parameter.

Σ_w::Array{Float64, 1}: Weights for each σ-point for the covariance.

m_w::Array{Float64, 1}: Weights for each σ-point for the mean.

residual_x::Function = -: Function to compute residual between states.

SigmaPoints.calculate_σ_points — Function

calculate_σ_points(
    ;σ_parameters::MerweScaled, x::Array{Float64, 1}, P::Array{Float64, 2}
)::Array{Float64, 2}

Compute 2n+1 x n Van der Merwe's σ-points for an Unscented Kalman Filter (UKF) given the mean x and covariance P.

Arguments

σ_parameters::MerweScaled: Parameters of the Van der Merwe's σ-points.

x::Array{Float64, 1}: Mean of the UKF.

P::Array{Float64, 2}: Covariance of the x of the UKF.

SigmaPoints.num_σ_points — Function

num_σ_points(σ_parameters::MerweScaled)::Int64

Return number of σ-points for given σ-parameters.