Unscented Kalman Filter

KalmanFilters.UnscentedKalmanFilter.UKFState — Type

UKFState(
    ;dim_x::Int64, dim_z::Int64, σ_parameters::MerweScaled,
    add_x::Function = default_add_x,
    residual_x::Function = default_residual,
    residual_z::Function = default_residual,
    mean_x::Function = default_mean_f,
    mean_z::Function = default_mean_f
)

State of the Unscented Kalman Filter.

Parameters

dim_x::Int64: State dimension

dim_z::Int64: Measurement dimension

x::Array{Float64, 1}: State of the filter

P::Array{Float64, 2}: Covariance matrix of the filter.

K::Array{Float64, 2}: Kalman gain of the filter.

S::Array{Float64, 2}: System uncertainty.

S_inv::Array{Float64, 2}: Inverse of system uncertainty.

σ_parameters::MerweScaled: Parameters for Van der Merwe sigma points.

σ_fx::Array{Float64, 2}: Sigma points after fx.

σ_hx::Array{Float64, 2}: Sigma points after hx.

add_x::Function: Function to calculate addition between two states.

residual_x::Function: Function to calculate residual between two states.

residual_z::Function: Function to calculate residual between state and measurement.

mean_x::Function: Function to calculate mean between Σ-points and weights after passing them through fx.

mean_z::Function: Function to calculate mean between Σ-points and weights after passing them through hx.

KalmanFilters.UnscentedKalmanFilter.predict! — Method

predict!(
    ;ukf::UKFState, δt::Float64, Q::Union{Nothing, Array{Float64, 2}},
    fx::Function, fx_args...
)

Perform prediction of the Unscented Kalman Filter (UKF). This updates x and P of the ukf to contain predicted state and covariance.

Arguments

ukf::UKFState: State of the UKF on which to perform prediction.

δt::Float64: Time delta.

Q::Union{Nothing, Array{Float64, 2}}: Process noise.

fx::Function(;x::Array{Float64, 1}, δt::Float64, ...): State transition function. Should accept keyword-only arguments, with mandatory x and δt. All other arguments may serve as control input or provide additional info.

fx_args...: Arguments to pass to fx function. Must contain at least x and δt arguments.

KalmanFilters.UnscentedKalmanFilter.update! — Method

update!(
    ;ukf::UKFState, z::Array{Float64, 1}, R::Union{Nothing, Array{Float64, 2}},
    hx::Function, hx_args...
)

Perform update step of the Unscented Kalman Filter (UKF) given measurements. This updates x and P of the ukf_state to contain updated state and covariance.

Arguments

ukf::UKFState: State of the UKF on which to perform update step.

z::Array{Float64, 1}: Measurements.

R::Union{Nothing, Array{Float64, 2}}: Measurement noise.

hx::Function: Measurement function. Should accept keyword-only arguments with mandatory x argument for state. Given state x of UKF it must transform it to the respective measurement.

hx_args...: Arguments that will be passed to hx function. Should contain at least x argument.