Unscented Kalman Filter

KalmanFilters.UnscentedKalmanFilter.UKFState — Type

State of the Unscented Kalman Filter

dim_x::Int64: State dimension

dim_z::Int64: Measurement dimension

x::AbstractArray{Float64, 1}: State of the filter

P::AbstractArray{Float64, 2}: Covariance matrix of the filter.

K::AbstractArray{Float64, 2}: Kalman gain of the filter.

S::AbstractArray{Float64, 2}: System uncertainty.

S_inv::AbstractArray{Float64, 2}: Inverse of system uncertainty.

σ_parameters::MerweScaled: Parameters for Van der Merwe sigma points.

σ_fx::AbstractArray{Float64, 2}: Sigma points after fx.

σ_hx::AbstractArray{Float64, 2}: Sigma points after hx.

add_x::Function: Function to calculate addition between two states.

residual_x::Function: Function to calculate residual between two states.

residual_z::Function: Function to calculate residual between state and measurement.

mean_x::Function: Function to calculate mean between Σ-points and weights after passing them through fx.

mean_z::Function: Function to calculate mean between Σ-points and weights after passing them through hx.

KalmanFilters.UnscentedKalmanFilter.UKFState — Method

UKFState(
    ;dim_x::Int64, dim_z::Int64, σ_parameters::MerweScaled,
    add_x::Function = default_add_x,
    residual_x::Function = default_residual,
    residual_z::Function = default_residual,
    mean_x::Function = default_mean_f,
    mean_z::Function = default_mean_f,
)

Arguments

dim_x::Int64: State dimension

dim_z::Int64: Measurement dimension

σ_parameters::MerweScaled: Parameters for Van der Merwe sigma points.

add_x::Function: Function to calculate addition between two states. Should accept two positional arguments y and x both of type AbstractArray{Float64, 1} and return AbstractArray{Float64, 1} as well.

residual_x::Function: Function to calculate residual between two states. Should accept two positional arguments y and x, but should have methods both for AbstractArray{Float64, 1}, AbstractArray{Float64, 1} and for AbstractArray{Float64, 2}, AbstractArray{Float64, 1} types of arguments. In second case, residual will be calculated between Σ-points and state x and return matrix of the same shape as Σ.

residual_z::Function: Function to calculate residual between state and measurement. Same as for residual_x, you should define methods for both kinds of arguments.

mean_x::Function: Function to calculate mean between Σ-points and weights after passing them through fx. Arguments are positional.

mean_z::Function: Function to calculate mean between Σ-points and weights after passing them through hx. Arguments are positional.

Note

Do not forget to implement methods for both kinds of arguments for residual_x and/or residual_z as described above.

KalmanFilters.UnscentedKalmanFilter.predict! — Method

predict!(
    ;ukf::UKFState, δt::Float64, Q::Union{Nothing, AbstractArray{Float64, 2}},
    fx::Function, fx_args...
)

Perform prediction of the Unscented Kalman Filter (UKF). This updates x and P of the ukf to contain predicted state and covariance.

Arguments

ukf::UKFState: State of the UKF on which to perform prediction.

δt::Float64: Time delta.

Q::Union{Nothing, AbstractArray{Float64, 2}}: Process noise.

fx::Function(;x::AbstractArray{Float64, 1}, δt::Float64, ...): State transition function. Should accept keyword-only arguments, with mandatory x and δt. All other arguments may serve as control input or provide additional info.

fx_args...: Arguments to pass to fx function. Must contain at least x and δt arguments.

Note

Just to emphasize, fx should accept keyword-only arguments, with mandatory x and δt. All other arguments may serve as control input or provide additional info.

KalmanFilters.UnscentedKalmanFilter.update! — Method

update!(
    ;ukf::UKFState, z::AbstractArray{Float64, 1}, R::Union{Nothing, AbstractArray{Float64, 2}},
    hx::Function, hx_args...
)

Perform update step of the Unscented Kalman Filter (UKF) given measurements. This updates x and P of the ukf_state to contain updated state and covariance.

Arguments

ukf::UKFState: State of the UKF on which to perform update step.

z::AbstractArray{Float64, 1}: Measurements.

R::Union{Nothing, AbstractArray{Float64, 2}}: Measurement noise.

hx::Function: Measurement function. Should accept keyword-only arguments with mandatory x argument for state. Given state x of UKF it must transform it to the respective measurement.

hx_args...: Arguments that will be passed to hx function. Should contain at least x argument.

Note

Just to emphasize, hx should accept keyword-only arguments with mandatory x argument for state. Given state x of UKF it must transform it to the respective measurement.

Default functions

KalmanFilters.UnscentedKalmanFilter.Helpers.default_add_x — Method

Default addition function.

KalmanFilters.UnscentedKalmanFilter.Helpers.default_mean_f — Method

Default mean function.

KalmanFilters.UnscentedKalmanFilter.Helpers.default_residual — Method

Default residual function.